requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
Flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
lxml==5.3.0
MarkupSafe==3.0.2
openai==1.58.1
priority==2.0.0
pydantic==2.10.4
pydantic_core==2.27.2
PyPDF2==3.0.1
python-docx==1.1.2
Quart==0.19.9
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
